markers connecting to listings, loading screen, and navigation
